refactor(macro): split js and jsx macro into two packages
BREAKING!
