+++ jest.config.js
@@ -63,6 +63,7 @@ module.exports = {
         "<rootDir>/packages/conf",
         "<rootDir>/packages/loader",
         "<rootDir>/packages/macro",
+        "<rootDir>/packages/react-macro",
         "<rootDir>/packages/vite-plugin",
         "<rootDir>/packages/format-po",
         "<rootDir>/packages/format-po-gettext",
