+++ packages/react-macro/__typetests__/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "jsx": "react",
     "esModuleInterop": true,
     "skipLibCheck": true
   },
