+++ packages/react-macro/build.config.ts
@@ -1,0 +1,5 @@
+import { defineBuildConfig } from "unbuild"
+
+export default defineBuildConfig({
+  declaration: false,
+})
