+++ packages/react-macro/src/index.ts
@@ -1,0 +1,3 @@
+import macro from "@lingui/macro"
+
+export default macro
